server/scripts/requirements.txt
requests
Pillow
fpdf
beautifulsoup4
cloudscraper